- more direct native entry points.

+++ src/sound/i_music.h
@@ -50,7 +50,7 @@ void I_BuildMIDIMenuList (FOptionValues *);
 void I_UpdateMusic ();
 
 // Volume.
-void I_SetMusicVolume (float volume);
+void I_SetMusicVolume (double volume);
 
 // Registers a song handle to song data.
 class MusInfo;
